image_scale_transformer as a wrapper to fix the bug in off_policy_driver
