+++ alf/environments/suite_mario.py
@@ -38,6 +38,7 @@ def load(game,
          state=None,
          discount=1.0,
          wrap_with_process=True,
+         gym_env_wrappers=(),
          frame_skip=4,
          frame_stack=4,
          data_format='channels_last',
@@ -91,7 +92,7 @@ def load(game,
             env,
             discount=discount,
             max_episode_steps=max_episode_steps,
-            gym_env_wrappers=(),
+            gym_env_wrappers=gym_env_wrappers,
             env_wrappers=(),
             spec_dtype_map=spec_dtype_map,
             auto_reset=False)
