+++ alf/environments/wrappers.py
@@ -17,6 +17,7 @@ import collections
 import gin
 import gym
 import numpy as np
+import tensorflow as tf
 
 
 @gin.configurable
@@ -89,6 +90,10 @@ class FrameSkip(gym.Wrapper):
         super().__init__(env)
         self._skip = skip
 
+    def __getattr__(self, name):
+        """Forward all other calls to the base environment."""
+        return getattr(self.env, name)
+
     def step(self, action):
         obs = None
         accumulated_reward = 0
@@ -102,3 +107,48 @@ class FrameSkip(gym.Wrapper):
 
     def reset(self, **kwargs):
         return self.env.reset(**kwargs)
+
+
+@gin.configurable
+class ImageScaleTransformer(gym.Wrapper):
+    """
+    Scale uint8 image input to min and max (0->min, 255->max)
+
+    Note: it treats an observation with len(shape)==4 as image
+    Args:
+        observation (nested Tensor): observations
+        min (float): normalize minimum to this value
+        max (float): normalize maximum to this value
+    Returns:
+        Transfromed observation
+    """
+    def __init__(self, env, minmax=(-1., 1.)):
+        super().__init__(env)
+        self._minmax = minmax
+
+    def __getattr__(self, name):
+        """Forward all other calls to the base environment."""
+        return getattr(self.env, name)
+
+    def reset(self):
+        observation = self.env.reset()
+        return self._transform(observation)
+
+    def step(self, action):
+        observation, reward, done, info = self.env.step(action)
+        return self._transform(observation), reward, done, info
+
+    def _transform(self, observation):
+        def _transform_image(obs):
+            # tf_agent changes all gym.spaces.Box observation to tf.float32.
+            # See _spec_from_gym_space() in tf_agents/environments/gym_wrapper.py
+            min, max = self._minmax
+            if len(obs.shape) == 4:
+                if obs.dtype == tf.uint8:
+                    obs = tf.cast(obs, tf.float32)
+                return ((max - min) / 255.) * obs + min
+            else:
+                return obs
+
+        return tf.nest.map_structure(
+            _transform_image, observation)
