+++ alf/examples/ac_breakout.gin
@@ -3,6 +3,7 @@
 
 # environment config
 import tf_agents.environments.suite_atari
+import alf.environments.wrappers
 # From OpenAI gym wiki:
 # "v0 vs v4: v0 has repeat_action_probability of 0.25
 #  (meaning 25% of the time the previous action will be used instead of the new action),
@@ -17,7 +18,7 @@ create_environment.num_parallel_environments=512
 # See my PR: https://github.com/tensorflow/agents/pull/138
 AtariPreprocessing.frame_skip=4
 
-suite_atari.load.gym_env_wrappers=(@AtariPreprocessing, @FrameStack4)
+suite_atari.load.gym_env_wrappers=(@AtariPreprocessing, @FrameStack4, @ImageScaleTransformer)
 
 # algorithm config
 ActorCriticLoss.entropy_regularization=0.01
@@ -50,5 +51,4 @@ on_policy_trainer.train.summary_interval=10
 on_policy_trainer.train.train_interval=3
 on_policy_trainer.train.use_tf_functions=1
 
-PolicyDriver.observation_transformer=@image_scale_transformer
 train_eval.debug_summaries=1
